K means 2.0 Mapping 1.0 Test set 1.0

+++ src/ch/epfl/advdb/milestone2/io/FVectorIMDB.java
@@ -19,7 +19,7 @@ public class FVectorIMDB extends FVector{
 		super(movieID);
 	}
 
-	public FVectorIMDB(Collection<? extends Integer> c, int movieID) {
+	public FVectorIMDB(Collection<? extends Float> c, int movieID) {
 		super(c, movieID);
 	}
 
@@ -35,22 +35,22 @@ public class FVectorIMDB extends FVector{
 	 * @see ch.epfl.advdb.milestone2.io.FVector#getDistance(ch.epfl.advdb.milestone2.io.ClusterCenter)
 	 */
 	@Override
-	public double getDistance(ClusterCenter c){
-		double sum=0;
-		double nv = 0;
-		double nc=0;
-		for (int index : this){
-			if(c.get(index)!=null)
+	public float getDistance(ClusterCenter c){
+		float sum=0;
+		float nv = 0;
+		float nc=0;
+		for (float index : this){
+			if(c.get((int)index)!=null)
 				//numerator
-				sum+=c.get(index);
+				sum+=c.get((int)index);
 			//sum Ai squared
 			nv++;
 		}
 		//sum Bi squared
-		for(double val : c.values())
+		for(float val : c.values())
 			nc+=val*val;
 		if(nc==0||nv==0)
 			return 0;
-		else return 1-(sum/(Math.sqrt(nv)*Math.sqrt(nc)));
+		else return (float) (1-(sum/(Math.sqrt(nv)*Math.sqrt(nc))));
 	}
 }
